requirements.txt
deepdiff==6.7.1
python-dotenv==1.0.1
Requests==2.31.0
requests_oauthlib==1.3.1
